librbd: Extract Group module from librbd/internal
